@@ -9,7 +9,7 @@ import re
 import uuid
 from decimal import Decimal
 from enum import Enum, IntEnum
-from typing import List, Union
+from typing import Union
 
 import pytz
 from pydantic import BaseModel, ConfigDict
@@ -960,7 +960,7 @@ class OldStyleModel(Model):
 
 
 def camelize_var(var_name: str):
-    var_parts: List[str] = var_name.split("_")
+    var_parts: list[str] = var_name.split("_")
     return var_parts[0] + "".join([part.title() for part in var_parts[1:]])
 
 
