@@ -1,17 +1,8 @@
 import asyncio
+from collections.abc import Callable, Coroutine
 from functools import wraps
 from itertools import count
-from typing import (
-    Any,
-    Callable,
-    Coroutine,
-    List,
-    Optional,
-    SupportsInt,
-    Tuple,
-    TypeVar,
-    Union,
-)
+from typing import Any, Optional, SupportsInt, TypeVar, Union
 
 try:
     import asyncmy as mysql
@@ -210,7 +201,7 @@ class MySQLClient(BaseDBAsyncClient):
     @translate_exceptions
     async def execute_query(
         self, query: str, values: Optional[list] = None
-    ) -> Tuple[int, List[dict]]:
+    ) -> tuple[int, list[dict]]:
         async with self.acquire_connection() as connection:
             self.log.debug("%s: %s", query, values)
             async with connection.cursor() as cursor:
@@ -221,7 +212,7 @@ class MySQLClient(BaseDBAsyncClient):
                     return cursor.rowcount, [dict(zip(fields, row)) for row in rows]
                 return cursor.rowcount, []
 
-    async def execute_query_dict(self, query: str, values: Optional[list] = None) -> List[dict]:
+    async def execute_query_dict(self, query: str, values: Optional[list] = None) -> list[dict]:
         return (await self.execute_query(query, values))[1]
 
     @translate_exceptions
