@@ -1,4 +1,6 @@
-from typing import TYPE_CHECKING, Any, List, Optional, Type
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any, Type
 
 from tortoise.backends.base.schema_generator import BaseSchemaGenerator
 from tortoise.converters import encoders
@@ -32,15 +34,15 @@ class OracleSchemaGenerator(BaseSchemaGenerator):
 
     def __init__(self, client: "OracleClient") -> None:
         super().__init__(client)
-        self._field_indexes: List[str] = []
-        self._foreign_keys: List[str] = []
-        self.comments_array: List[str] = []
+        self._field_indexes: list[str] = []
+        self._foreign_keys: list[str] = []
+        self.comments_array: list[str] = []
 
     def quote(self, val: str) -> str:
         return f'"{val}"'
 
     @classmethod
-    def _get_escape_translation_table(cls) -> List[str]:
+    def _get_escape_translation_table(cls) -> list[str]:
         table = super()._get_escape_translation_table()
         table[ord("'")] = "''"
         return table
@@ -88,18 +90,18 @@ class OracleSchemaGenerator(BaseSchemaGenerator):
     def _get_index_sql(
         self,
         model: "Type[Model]",
-        field_names: List[str],
+        field_names: list[str],
         safe: bool,
-        index_name: Optional[str] = None,
-        index_type: Optional[str] = None,
-        extra: Optional[str] = None,
+        index_name: str | None = None,
+        index_type: str | None = None,
+        extra: str | None = None,
     ) -> str:
-        return super(OracleSchemaGenerator, self)._get_index_sql(
+        return super()._get_index_sql(
             model, field_names, False, index_name=index_name, index_type=index_type, extra=extra
         )
 
     def _get_table_sql(self, model: "Type[Model]", safe: bool = True) -> dict:
-        return super(OracleSchemaGenerator, self)._get_table_sql(model, False)
+        return super()._get_table_sql(model, False)
 
     def _create_fk_string(
         self,
