@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-from typing import Optional
-
 from pypika_tortoise import Parameter
 
 from tortoise import Model
@@ -9,9 +7,7 @@ from tortoise.backends.base_postgres.executor import BasePostgresExecutor
 
 
 class PsycopgExecutor(BasePostgresExecutor):
-    async def _process_insert_result(
-        self, instance: Model, results: Optional[dict | tuple]
-    ) -> None:
+    async def _process_insert_result(self, instance: Model, results: dict | tuple | None) -> None:
         if results:
             db_projection = instance._meta.fields_db_projection_reverse
 
