@@ -1,5 +1,6 @@
+from collections.abc import Iterable
 from types import ModuleType
-from typing import Dict, Iterable, Optional, Union
+from typing import Optional, Union
 
 from blacksheep import Request
 from blacksheep.server import Application
@@ -15,7 +16,7 @@ def register_tortoise(
     config: Optional[dict] = None,
     config_file: Optional[str] = None,
     db_url: Optional[str] = None,
-    modules: Optional[Dict[str, Iterable[Union[str, ModuleType]]]] = None,
+    modules: Optional[dict[str, Iterable[Union[str, ModuleType]]]] = None,
     generate_schemas: bool = False,
     add_exception_handlers: bool = False,
 ) -> None:
