@@ -1,7 +1,8 @@
 import asyncio
 import logging
+from collections.abc import Iterable
 from types import ModuleType
-from typing import Dict, Iterable, Optional, Union
+from typing import Optional, Union
 
 from quart import Quart  # pylint: disable=E0401
 
@@ -14,7 +15,7 @@ def register_tortoise(
     config: Optional[dict] = None,
     config_file: Optional[str] = None,
     db_url: Optional[str] = None,
-    modules: Optional[Dict[str, Iterable[Union[str, ModuleType]]]] = None,
+    modules: Optional[dict[str, Iterable[Union[str, ModuleType]]]] = None,
     generate_schemas: bool = False,
 ) -> None:
     """
