@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Callable, Type
+from collections.abc import Callable
+from typing import TYPE_CHECKING, Any, Type
 
 from tortoise.connection import connections
 from tortoise.exceptions import ConfigurationError
