@@ -1,8 +1,9 @@
 import datetime
 import time
+from collections.abc import Sequence
 from datetime import timedelta
 from decimal import Decimal
-from typing import Any, Dict, Sequence, Set
+from typing import Any
 
 _escape_table = [chr(x) for x in range(128)]
 _escape_table[0] = "\\0"
@@ -45,7 +46,7 @@ def escape_item(val: Any, mapping=None) -> str:
     return val
 
 
-def escape_dict(val: Dict, mapping=None) -> dict:
+def escape_dict(val: dict, mapping=None) -> dict:
     n = {}
     for k, v in val.items():
         quoted = escape_item(v, mapping)
@@ -61,7 +62,7 @@ def escape_sequence(val: Sequence, mapping=None) -> str:
     return "'{" + ",".join(n) + "}'"
 
 
-def escape_set(val: Set, mapping=None) -> str:
+def escape_set(val: set, mapping=None) -> str:
     return ",".join([escape_item(x, mapping) for x in val])
 
 
