@@ -1,10 +1,11 @@
 from __future__ import annotations
 
 import operator
+from collections.abc import Iterator
 from dataclasses import dataclass
 from dataclasses import field as dataclass_field
 from enum import Enum
-from typing import TYPE_CHECKING, Any, Iterator, Type, cast
+from typing import TYPE_CHECKING, Any, Type, cast
 
 from pypika_tortoise import Case as PypikaCase
 from pypika_tortoise import Field as PypikaField
